Remove dead code path and update snapshot.
This path handled top statements such as:

```php
namespace\foo();
```

This would end up in `read_namespace` due to the `T_NAMESPACE` token. However,
with the changes in the previous two commits this is now parsed as a single
`T_NAME_RELATIVE` token, and is parsed as a regular call expression statement
instead.